Update Rotation Matching.cpp

+++ 1400/Rotation Matching.cpp
@@ -8,7 +8,7 @@ int main()
 	ll n;
 	cin>>n;
 	vector<ll>a(n+1);
-	vector<int>b(n+1);
+	vector<ll>b(n+1);
 	vector<int>p(n+1);
 	map<int,int>m;
 	for(int i=1;i<=n;i++)
